Refactor code to compare and plot data for gl880
This commit refactors the code in `compare_data.py` to compare and plot data for the target `gl880`. It imports necessary modules, retrieves the path, and loads the data files. It then plots the flux for each order and calculates the signal-to-noise ratio (SNR) for each order. The code also updates the `config_freechem.py` file by changing the value of the `run` variable to 'sphinx16' and updating the `file_target` parameter to 'data/spec_orders_46_47_48_mad.npy'. Additionally, the `N_knots` parameter is changed to 25 in the `config_freechem.py` file. Finally, the `free_parameter_check.py` and `retrieval_script.py` files are modified to update the `run` variable to 'sphinx16'. The changes in the `retrieval.py` file include updating the `prior_check` function to generate random values between 0 and 1 for the `theta` parameter.